scripts/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
duckdb==1.1.1